requirements.txt
flask
requests
beautifulsoup4
openai
pinecone-client
flask-cors
python-dotenv
sentence-transformers
torch
transformers
